chore: Rename lbforaging to sustainable_foraging, update pyproject.toml
