+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include lbforaging/foraging/icons/*.png
+include sustainable_foraging/foraging/icons/*.png
