--- lbforaging/agents/__init__.py
@@ -1,6 +1,0 @@
-from lbforaging.agents.agent import BaseAgent  # noqa
-from lbforaging.agents.random_agent import RandomAgent  # noqa
-from lbforaging.agents.heuristic_agent import H1, H2, H3, H4  # noqa
-from lbforaging.agents.q_agent import QAgent  # noqa
-from lbforaging.agents.monte_carlo import MonteCarloAgent  # noqa
-from lbforaging.agents.hba import HBAAgent  # noqa
