+++ scripts/inference_sb3.py
@@ -3,8 +3,8 @@ import time
 
 import numpy as np
 import supersuit as ss
-from lbforaging.foraging import AECForagingEnv
-from lbforaging.foraging.sustainable_benchmark import (
+from sustainable_foraging.foraging import AECForagingEnv
+from sustainable_foraging.foraging.sustainable_benchmark import (
     BENCHMARK_NAME,
     get_preset,
     list_presets,
@@ -21,7 +21,7 @@ def main():
         "--model",
         "-m",
         type=str,
-        default="ppo_lbforaging",
+        default="ppo_sustainable_foraging",
         help="Path to the trained model zip file (without .zip extension)",
     )
     parser.add_argument(
