+++ scripts/train_sb3_a2c.py
@@ -18,8 +18,8 @@ import supersuit as ss
 from stable_baselines3 import A2C
 from stable_baselines3.common.callbacks import BaseCallback
 
-from lbforaging.foraging import AECForagingEnv
-from lbforaging.foraging.sustainable_benchmark import get_preset, get_training_defaults
+from sustainable_foraging.foraging import AECForagingEnv
+from sustainable_foraging.foraging.sustainable_benchmark import get_preset, get_training_defaults
 from pettingzoo.utils import aec_to_parallel
 from scripts._bench_utils import MetricsTracker, get_standard_parser, save_experiment_config
