+++ sustainable_foraging/agents/__init__.py
@@ -1,0 +1,6 @@
+from sustainable_foraging.agents.agent import BaseAgent  # noqa
+from sustainable_foraging.agents.random_agent import RandomAgent  # noqa
+from sustainable_foraging.agents.heuristic_agent import H1, H2, H3, H4  # noqa
+from sustainable_foraging.agents.q_agent import QAgent  # noqa
+from sustainable_foraging.agents.monte_carlo import MonteCarloAgent  # noqa
+from sustainable_foraging.agents.hba import HBAAgent  # noqa
